Revert "Refact/main page"
