+++ src/App.js
@@ -1,9 +1,11 @@
 import React, { Component } from 'react';
+
 import Sidebar from 'react-sidebar';
 import Header from './components/Header';
-import SCSidebar from './components/Sidebar';
-import GlobalStyle from './globalStyle';
 import Routes from './Routes';
+import GlobalStyle from './globalStyle';
+
+import SCSidebar from './components/Sidebar';
 
 class App extends Component {
     constructor(props) {
