+++ src/components/shared/table/Table.js
@@ -5,40 +5,36 @@ import utils from '../../utils';
 
 const Table = ({ headers, rows, imageRows, color, children }) => (
     <SCTable color={color}>
-        <table>
-            <thead>
-                <tr>
-                    {headers.map(header => (
-                        <th key={header}>{header}</th>
+        <thead>
+            <tr>
+                {headers.map(header => (
+                    <th key={header}>{header}</th>
+                ))}
+            </tr>
+        </thead>
+        <tbody>
+            {children}
+            <tr>
+                {imageRows &&
+                    imageRows.map((row, index) => (
+                        <td key={index + row}>
+                            {utils.getImage(row.types, row.wantedType) && (
+                                <Image
+                                    imageUrl={utils.getImage(
+                                        row.types,
+                                        row.wantedType
+                                    )}
+                                    altText={row.yokaiType}
+                                    size={row.size}
+                                />
+                            )}
+                        </td>
                     ))}
-                </tr>
-            </thead>
-            <tbody>
-                {children}
-                <tr>
-                    {imageRows &&
-                        imageRows.map((row, index) => (
-                            <td key={index + row}>
-                                {utils.getImage(row.types, row.wantedType) && (
-                                    <Image
-                                        imageUrl={utils.getImage(
-                                            row.types,
-                                            row.wantedType
-                                        )}
-                                        altText={row.yokaiType}
-                                        size={row.size}
-                                    />
-                                )}
-                            </td>
-                        ))}
 
-                    {rows &&
-                        rows.map((row, index) => (
-                            <td key={index + row}>{row}</td>
-                        ))}
-                </tr>
-            </tbody>
-        </table>
+                {rows &&
+                    rows.map((row, index) => <td key={index + row}>{row}</td>)}
+            </tr>
+        </tbody>
     </SCTable>
 );
