--- src/components/shared/variables/Yokais.js
@@ -1,25 +1,0 @@
-export const Ranks = ['A', 'B', 'C', 'D', 'E', 'S'];
-
-export const TribesName = [
-    'Brave',
-    'Charming',
-    'Eerie',
-    'Heartful',
-    'Mysterious',
-    'Shady',
-    'Slippery',
-    'Tough',
-    'Wicked',
-    'Boss'
-];
-
-export const Elements = [
-    'Drain',
-    'Earth',
-    'Fire',
-    'Ice',
-    'Lightning',
-    'Restoration',
-    'Water',
-    'Wind'
-];
